+++ node/status_node_services.go
@@ -40,7 +40,7 @@ import (
 	appmetricsservice "github.com/status-im/status-go/services/appmetrics"
 	"github.com/status-im/status-go/services/browsers"
 	"github.com/status-im/status-go/services/chat"
-	"github.com/status-im/status-go/services/communitytokensv2"
+	"github.com/status-im/status-go/services/communitytokens"
 	"github.com/status-im/status-go/services/connector"
 	"github.com/status-im/status-go/services/ens"
 	"github.com/status-im/status-go/services/eth"
@@ -96,7 +96,7 @@ func (b *StatusNode) initServices(config *params.NodeConfig, mediaServer *server
 	services = append(services, b.statusPublicService())
 	services = append(services, b.pendingTrackerService(&b.walletFeed))
 	services = append(services, b.ensService(b.timeSourceNow()))
-	services = append(services, b.CommunityTokensServiceV2())
+	services = append(services, b.CommunityTokensService())
 	services = append(services, b.stickersService(accDB))
 	services = append(services, b.updatesService())
 	services = appendIf(b.appDB != nil && b.multiaccountsDB != nil, services, b.accountsService(&b.accountsFeed, accDB, mediaServer))
@@ -502,11 +502,11 @@ func (b *StatusNode) pendingTrackerService(walletFeed *event.Feed) *transactions
 	return b.pendingTracker
 }
 
-func (b *StatusNode) CommunityTokensServiceV2() *communitytokensv2.Service {
-	if b.communityTokensSrvcV2 == nil {
-		b.communityTokensSrvcV2 = communitytokensv2.NewService(b.rpcClient, b.gethAccountManager, b.config, b.appDB, &b.walletFeed, b.transactor)
+func (b *StatusNode) CommunityTokensService() *communitytokens.Service {
+	if b.communityTokensSrvc == nil {
+		b.communityTokensSrvc = communitytokens.NewService(b.rpcClient, b.gethAccountManager, b.config, b.appDB, &b.walletFeed, b.transactor)
 	}
-	return b.communityTokensSrvcV2
+	return b.communityTokensSrvc
 }
 
 func (b *StatusNode) stickersService(accountDB *accounts.Database) *stickers.Service {
