+++ services/communitytokens/api_test.go
@@ -1,4 +1,4 @@
-package communitytokensv2
+package communitytokens
 
 import (
 	"testing"
